+++ packages/plugin/src/plugin/url.ts
@@ -1,9 +1,9 @@
-import path from 'path';
+import path from 'node:path';
 import { JSONOutput, ReflectionKind } from 'typedoc';
 import type { ResolvedPackageConfig } from '../types';
 
 export function joinUrl(...paths: string[]): string {
-	return path.join(...paths).replace(/\\/g, '/');
+	return path.join(...paths).replaceAll('\\', '/');
 }
 
 export function getKindSlug(decl: JSONOutput.DeclarationReflection): string {
@@ -36,14 +36,17 @@ export function getPackageSlug(
 	}
 
 	// packages/foo -> foo
-	const slug = pkgConfig.packageSlug ?? path.basename(pkgConfig.packagePath);
-
-	// bar/baz -> bar-baz
-	const importName = importPath.replace(/\\/g, '-');
-
-	if (importName === 'index') {
-		return slug;
+	let slug = pkgConfig.packageSlug ?? path.basename(pkgConfig.packagePath);
+	if (slug === '.') {
+		slug = '';
 	}
 
-	return `${slug}-${importName}`;
+	// bar/baz -> bar-baz
+	const importName = importPath.replaceAll('\\', '-');
+
+	if (importName === 'index') {
+		return slug || '.';
+	}
+
+	return slug ? `${slug}-${importName}` : importName;
 }