+++ packages/plugin/src/types.ts
@@ -150,6 +150,6 @@ export interface TSDSignatureReflection extends JSONOutput.SignatureReflection {
 }
 
 declare global {
-	// eslint-disable-next-line no-var, vars-on-top
+	 
 	var typedocBuild: { count: number };
 }